ZNTA-1865 - refactor move methods and class
