+++ server/zanata-war/src/main/java/org/zanata/service/TransMemoryMergeService.java
@@ -29,6 +29,7 @@ import org.zanata.async.handle.TransMemoryMergeTaskHandle;
 import org.zanata.model.HLocale;
 import org.zanata.model.HProjectIteration;
 import org.zanata.model.HTextFlow;
+import org.zanata.rest.dto.VersionTMMerge;
 import org.zanata.webtrans.shared.model.ProjectIterationId;
 import org.zanata.webtrans.shared.rest.dto.HasTMMergeCriteria;
 import org.zanata.webtrans.shared.rest.dto.TransMemoryMergeRequest;
@@ -39,10 +40,27 @@ public interface TransMemoryMergeService extends Serializable {
             TransMemoryMergeRequest request,
             TransMemoryMergeTaskHandle asyncTaskHandle);
 
+    /**
+     * TM merge for a single document
+     * @param request
+     * @param asyncTaskHandle
+     * @return
+     */
     Future<List<TranslationService.TranslationResult>> executeMergeAsync(TransMemoryMergeRequest request,
             TransMemoryMergeTaskHandle asyncTaskHandle);
 
     void translateInBatch(HasTMMergeCriteria mergeCriteria,
             List<HTextFlow> textFlows,
             HLocale targetLocale, List<Long> fromProjectVersionIds);
+
+    /**
+     * TM merge for a project version
+     * @param versionId
+     * @param mergeRequest
+     * @param handle
+     * @return
+     */
+    Future<Void> startMergeTranslations(Long versionId,
+            VersionTMMerge mergeRequest,
+            MergeTranslationsTaskHandle handle);
 }
